#64, #69, #66, #70 Implement Follow Request Sending Functionality
Took some inspiration from Animesh's UI and streamlined it then added all the Stuff to it dded a followers and following and its like insta where u just press on it and u can see ur followers or following and u can even search for users and stuff on the following and followings. You can also remove followers and it will give u a pop up to make sure. Then I also added a Follow Requests page u just press on it and u can accept and decline. And if u do accept it it gives an option to follow back (just like insta or snap). Theres also the search Users which then again u can search a user and then send a request. The thing makes sure that if u do follow them already and u do search for them it shows up as following in the search (just like insta or snap).  Settings has the logout option and the Home page is kinda set up like insta where it has all the followings recent posts and it only shows the latests weekly post of each the followings.