#64, #69, #66, #70 Implement Follow Request Sending Functionality (#160)
* #64, #69, #66, #70 Implement Follow Request Sending Functionality

Took some inspiration from Animesh's UI and streamlined it then added all the Stuff to it dded a followers and following and its like insta where u just press on it and u can see ur followers or following and u can even search for users and stuff on the following and followings. You can also remove followers and it will give u a pop up to make sure. Then I also added a Follow Requests page u just press on it and u can accept and decline. And if u do accept it it gives an option to follow back (just like insta or snap). Theres also the search Users which then again u can search a user and then send a request. The thing makes sure that if u do follow them already and u do search for them it shows up as following in the search (just like insta or snap).  Settings has the logout option and the Home page is kinda set up like insta where it has all the followings recent posts and it only shows the latests weekly post of each the followings.

* update

Just resolved minor UI disrepencies and cleaned up the code

* Had to fix query structure

removed the orderBy("timestamp", Query.Direction.DESCENDING), Instead of asking Firestore to sort the data, now retrieving the unordered data and sorting it ourselves in app's memory after receiving it.

* Made Necessary via comments

* update regarding the comments

* fixed some firestore index errors

* Changes

Added back the search feature in the homefragment, and then fixed all the comments and javadocs and then fixed the indexing in the moodeventrepository.

* fixed all UI problems and Test

Fixed the Test completly and it passees, Fixed the rest of the fragments since the UI was not sitting well.

* Had to fix a bug with the Follow back feature

Also got rid of the delete feature in the previous commit.

* final (hopfeully)

fixed these issues : Added the usuage of TimestamUtils on pfp, and then MoodEvent Chip got fixed. And reverted back to old layout_event_home